+++ webapp/src/main/java/ar/edu/itba/paw/webapp/dto/AssetsInstancesDTO.java
@@ -32,7 +32,7 @@ public class AssetsInstancesDTO {
 
     private String imageReference;
 
-    private Integer rating;
+    private float rating;
 
     private String selfUrl;
